size_t -> int

+++ src/base/symbolic_execution.cc
@@ -52,7 +52,7 @@ void SymbolicExecution::Serialize(string* s) const {
 
 bool SymbolicExecution::Parse(istream& s) {
   // Read the inputs.
-  size_t len;
+  int len;
   char buf[MAX_LINE_BUF];
 
   s.getline(buf, MAX_LINE_BUF);
@@ -63,7 +63,7 @@ bool SymbolicExecution::Parse(istream& s) {
   vars_.clear();
   inputs_.resize(len);
 
-  for (size_t i = 0; i < len; i++) {
+  for (int i = 0; i < len; i++) {
     int type;
     long long int value;
     s.getline(buf, MAX_LINE_BUF);
